use SecurityManager, refactor code
